rename http_request to request according to the class naming
